+++ src/views/settings/components/Http.vue
@@ -1,24 +1,24 @@
 <template>
   <div v-if="!loading" :class="isSidebarOpen" class="form-container">
-    <el-form ref="httpData" :model="httpData" :label-width="labelWidth">
+    <el-form ref="httpData" :model="httpData" :label-position="labelPosition" :label-width="labelWidth">
       <setting :setting-group="http" :data="httpData"/>
     </el-form>
-    <el-form ref="corsPlugData" :model="corsPlugData" :label-width="labelWidth">
+    <el-form ref="corsPlugData" :model="corsPlugData" :label-position="labelPosition" :label-width="labelWidth">
       <el-form-item data-search=":cors_plug" class="grouped-settings-header">
-        <span class="label-font">{{ $t('settings.corsPlug') }}</span>
+        <span class="label-font label-with-margin">{{ $t('settings.corsPlug') }}</span>
       </el-form-item>
       <setting :setting-group="corsPlug" :data="corsPlugData"/>
     </el-form>
     <el-divider class="divider thick-line"/>
-    <el-form ref="httpSignatures" :model="httpSignaturesData" :label-width="labelWidth">
+    <el-form ref="httpSignatures" :model="httpSignaturesData" :label-position="labelPosition" :label-width="labelWidth">
       <setting :setting-group="httpSignatures" :data="httpSignaturesData"/>
     </el-form>
     <el-divider class="divider thick-line"/>
-    <el-form ref="httpSecurityData" :model="httpSecurityData" :label-width="labelWidth">
+    <el-form ref="httpSecurityData" :model="httpSecurityData" :label-position="labelPosition" :label-width="labelWidth">
       <setting :setting-group="httpSecurity" :data="httpSecurityData"/>
     </el-form>
     <el-divider class="divider thick-line"/>
-    <el-form ref="webCacheTtl" :model="webCacheTtlData" :label-width="labelWidth">
+    <el-form ref="webCacheTtl" :model="webCacheTtlData" :label-position="labelPosition" :label-width="labelWidth">
       <setting :setting-group="webCacheTtl" :data="webCacheTtlData"/>
     </el-form>
     <div class="submit-button-container">
@@ -72,6 +72,9 @@ export default {
     },
     isTablet() {
       return this.$store.state.app.device === 'tablet'
+    },
+    labelPosition() {
+      return this.isMobile ? 'top' : 'right'
     },
     labelWidth() {
       if (this.isMobile) {